+++ src/io/format.rs
@@ -15,6 +15,9 @@ pub enum ImageFormat {
     /// An Image in JPEG Format
     Jpeg,
 
+    /// An Image in JPEG XL Format
+    JpegXl,
+
     /// An Image in GIF Format
     Gif,
 
@@ -75,6 +78,7 @@ impl ImageFormat {
             Some(match ext.as_str() {
                 "avif" => ImageFormat::Avif,
                 "jpg" | "jpeg" | "jfif" => ImageFormat::Jpeg,
+                "jxl" => ImageFormat::JpegXl,
                 "png" | "apng" => ImageFormat::Png,
                 "gif" => ImageFormat::Gif,
                 "webp" => ImageFormat::WebP,
@@ -151,6 +155,7 @@ impl ImageFormat {
         match mime_type.as_str() {
             "image/avif" => Some(ImageFormat::Avif),
             "image/jpeg" => Some(ImageFormat::Jpeg),
+            "image/jxl" => Some(ImageFormat::JpegXl),
             "image/png" => Some(ImageFormat::Png),
             "image/gif" => Some(ImageFormat::Gif),
             "image/webp" => Some(ImageFormat::WebP),
@@ -196,6 +201,7 @@ impl ImageFormat {
         match self {
             ImageFormat::Avif => "image/avif",
             ImageFormat::Jpeg => "image/jpeg",
+            ImageFormat::JpegXl => "image/jxl",
             ImageFormat::Png => "image/png",
             ImageFormat::Gif => "image/gif",
             ImageFormat::WebP => "image/webp",
@@ -231,6 +237,7 @@ impl ImageFormat {
         match self {
             ImageFormat::Png => &["png"],
             ImageFormat::Jpeg => &["jpg", "jpeg"],
+            ImageFormat::JpegXl => &["jxl"],
             ImageFormat::Gif => &["gif"],
             ImageFormat::WebP => &["webp"],
             ImageFormat::Pnm => &["pbm", "pam", "ppm", "pgm", "pnm"],
@@ -258,6 +265,7 @@ impl ImageFormat {
             ImageFormat::Png => cfg!(feature = "png"),
             ImageFormat::Gif => cfg!(feature = "gif"),
             ImageFormat::Jpeg => cfg!(feature = "jpeg"),
+            ImageFormat::JpegXl => cfg!(feature = "jxl"),
             ImageFormat::WebP => cfg!(feature = "webp"),
             ImageFormat::Tiff => cfg!(feature = "tiff"),
             ImageFormat::Tga => cfg!(feature = "tga"),
@@ -283,6 +291,7 @@ impl ImageFormat {
             ImageFormat::Gif => cfg!(feature = "gif"),
             ImageFormat::Ico => cfg!(feature = "ico"),
             ImageFormat::Jpeg => cfg!(feature = "jpeg"),
+            ImageFormat::JpegXl => false,
             ImageFormat::Png => cfg!(feature = "png"),
             ImageFormat::Bmp => cfg!(feature = "bmp"),
             ImageFormat::Tiff => cfg!(feature = "tiff"),
@@ -303,6 +312,7 @@ impl ImageFormat {
             ImageFormat::Gif,
             ImageFormat::Ico,
             ImageFormat::Jpeg,
+            ImageFormat::JpegXl,
             ImageFormat::Png,
             ImageFormat::Bmp,
             ImageFormat::Tiff,
@@ -335,6 +345,7 @@ mod tests {
         assert_eq!(from_path("./a.jpg").unwrap(), ImageFormat::Jpeg);
         assert_eq!(from_path("./a.jpeg").unwrap(), ImageFormat::Jpeg);
         assert_eq!(from_path("./a.JPEG").unwrap(), ImageFormat::Jpeg);
+        assert_eq!(from_path("./a.jxl").unwrap(), ImageFormat::JpegXl);
         assert_eq!(from_path("./a.pNg").unwrap(), ImageFormat::Png);
         assert_eq!(from_path("./a.gif").unwrap(), ImageFormat::Gif);
         assert_eq!(from_path("./a.webp").unwrap(), ImageFormat::WebP);
@@ -358,7 +369,7 @@ mod tests {
     fn image_formats_are_recognized() {
         use ImageFormat::*;
         const ALL_FORMATS: &[ImageFormat] = &[
-            Avif, Png, Jpeg, Gif, WebP, Pnm, Tiff, Tga, Bmp, Ico, Hdr, Farbfeld, OpenExr,
+            Avif, Png, Jpeg, JpegXl, Gif, WebP, Pnm, Tiff, Tga, Bmp, Ico, Hdr, Farbfeld, OpenExr,
         ];
         for &format in ALL_FORMATS {
             let mut file = Path::new("file.nothing").to_owned();
@@ -409,6 +420,7 @@ mod tests {
         assert!(all_formats.contains(&ImageFormat::Bmp));
         assert!(all_formats.contains(&ImageFormat::Farbfeld));
         assert!(all_formats.contains(&ImageFormat::Jpeg));
+        assert!(all_formats.contains(&ImageFormat::JpegXl));
     }
 
     #[test]